update Waggy version and show example of using a WaggyRouter
