+++ challenge-1/go.mod
@@ -4,4 +4,4 @@ go 1.17
 
 require github.com/fermyon/spin/sdk/go v0.6.0
 
-require github.com/syke99/waggy v0.4.13
+require github.com/syke99/waggy v0.4.14
